+++ packages/web3-types/src/web3_api_types.ts
@@ -17,6 +17,24 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 
 import { JsonRpcId, JsonRpcIdentifier } from './json_rpc_types';
 
+export interface ProviderMessage {
+	readonly type: string;
+	readonly data: unknown;
+}
+export interface EthSubscription extends ProviderMessage {
+	readonly type: 'eth_subscription';
+	readonly data: {
+		readonly subscription: string;
+		readonly result: unknown;
+	};
+}
+export interface ProviderRpcError extends Error {
+	code: number;
+	data?: unknown;
+}
+export interface ProviderConnectInfo {
+	readonly chainId: string;
+}
 export type Web3APISpec = Record<string, (...params: any) => any> | unknown;
 export type Web3APIMethod<T extends Web3APISpec> = string & keyof Exclude<T, unknown>;
 export type Web3APIParams<